Remove cell number and language from pod
We started using the identity store and the hub to fetch this information,
but unfortunately the field names are different depending on which service
the info is coming from.

These 2 fields are already displayed in the CasePro interface so it makes
sense to not use the pod at all for them.